Move all blockly scripts together & rename
