--- src/components/challengeView/BlocksWorkspace.tsx
@@ -1,9 +1,0 @@
-import { ToolboxPreview } from "../blockly/ToolboxPreview"
-
-export const BlocksWorkspace = () => {
-    return <ToolboxPreview
-            sx={{flexGrow:1}} 
-            blocksToPreview={["MoverACasillaDerecha"]}
-            categorized={false} 
-        />
-}
