+++ src/components/challengeView/EditableBlocklyWorkspace.tsx
@@ -1,0 +1,9 @@
+import { PBBlocklyWorkspace } from "../blockly/PBBlocklyWorkspace"
+
+export const EditableBlocklyWorkspace = () => {
+    return <PBBlocklyWorkspace
+            sx={{flexGrow:1}} 
+            blocksToPreview={["MoverACasillaDerecha"]}
+            categorized={false} 
+        />
+}
